# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Phober - Notification Server",
        "contact": {
            "name": "Hikmat",
            "url": "https://www.linkedin.com/in/abdukhaligov/",
            "email": "hikmat.pou@gmail.com"
        },
        "version": "V 1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "localhost"
        }
    ],
    "paths": {
        "/otp": {
            "get": {
                "tags": [
                    "OTP"
                ],
                "summary": "Send OTP",
                "operationId": "sendOtp",
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": [
                    {
                        "auth.server": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "OTP"
                ],
                "summary": "Check OTP",
                "operationId": "checkOtp",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "code": "A1B2C3",
                                "id": "1815ce18-0b4c-4d5f-ac73-a3b4aba36553"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": [
                    {
                        "auth.server": []
                    }
                ]
            }
        },
        "/support/message": {
            "post": {
                "tags": [
                    "Support"
                ],
                "summary": "Send message to support",
                "operationId": "supportMessage",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {},
                            "example": {
                                "subject": "Subject",
                                "message": "Message"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ok"
                    }
                },
                "security": [
                    {
                        "auth.server": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "auth.server": {
                "type": "http",
                "description": "Authorization token obtained from logging in.",
                "name": "Authorization",
                "in": "header",
                "scheme": "bearer"
            }
        }
    }
}